Move LR0Generator to a new module, jsparagus.lr0.
